Extract _named_config.py from _legacy.py; fix stale re-export shims
Moves save_to_named_config out of _legacy.py and into its own module.
This is the last function in _legacy.py — the file now contains only
constants (33 lines from 1613 at sprint start; 98% reduction).

New file: modules/helpers/_named_config.py (133 lines)
  - save_to_named_config: persist a generated YAML under a named identity,
    handling archive rotation, dual writes (Quickstart + Kometa config dirs),
    font sync, and managed library artifact sync. Behavior unchanged.
  - DRY tidy: consolidated the two duplicate 'from modules import helpers'
    lazy imports (previously aliased as both _h_artifacts and _helpers) into
    a single '_h' alias used throughout.

Fix stale re-export shims: several submodules were still importing symbols
'through' _legacy.py that had already moved. Ruff auto-pruned the unused
'from ._logging import ts_log' line at the top of _legacy.py, which broke:

  - _settings.py (booler): 'from modules.helpers._legacy import ts_log'
    -> now imports from _logging directly
  - _restart.py (set_restart_notice, consume_restart_notice): same fix
    (also consolidated to a single top-level import)
  - _artifacts.py (3 sites) and _fonts.py: 'from modules.helpers._legacy
    import get_kometa_config_dir' -> now import from _kometa_paths directly

Each symbol now has one source of truth (DRY).

External call sites unchanged: quickstart.py, blueprints/config_routes.py,
tests/test_core_backend.py all reach save_to_named_config via
helpers.save_to_named_config() — the star re-export in helpers/__init__.py
keeps them working.

Reduces _legacy.py from 157 to 33 lines (-124). All 1077 tests pass.
