+++ modules/helpers/__init__.py
@@ -51,6 +51,7 @@ from ._kometa_version import *  # noqa: F403
 from ._pid import *  # noqa: F403
 from ._paths import *  # noqa: F403
 from ._kometa_paths import *  # noqa: F403
+from ._named_config import *  # noqa: F403
 from ._zip_update import *  # noqa: F403
 from ._logging import *  # noqa: F403
 from ._plex import *  # noqa: F403
